+++ CHANGES.md
@@ -1,5 +1,6 @@
 ### Last Version
 
+  * Fix bug with calling assert\object with a non object and ask to be cloned
   * Fix $path in schemas called by assert\all
   * Parameter extra for assert\dict accepts a schema dict